edge-node/requirements.txt
opencv-python
opencv-contrib-python
numpy
Pillow
python-socketio==5.7.2
